feat: refactor remaining discount api
