+++ fwb/utils/discount_utils.ts
@@ -1,8 +1,13 @@
 import { NextRequest, NextResponse } from "next/server";
 import { auth, currentUser } from "@clerk/nextjs";
-import { supabaseClientWithAccessToken } from "@/utils/supabase_utils";
 import supabaseClient from "@/supabase";
 
+/**
+ * Inserts a new discount into the database.
+ * 
+ * @param {NextRequest} request - The request object.
+ * @returns {Promise<NextResponse>} A promise that resolves to the response object.
+ */
 const insertDiscount = async (request: NextRequest) => {
   try {
     const { userId, getToken } = auth();
@@ -23,7 +28,7 @@ const insertDiscount = async (request: NextRequest) => {
         private_groups: formData.get("private_groups"),
       };
 
-      const supabase = await supabaseClientWithAccessToken(
+      const supabase = await supabaseClient(
         request.headers.get("supabase_jwt")
       );
       if (!supabase) {
@@ -60,6 +65,12 @@ const insertDiscount = async (request: NextRequest) => {
   }
 };
 
+/**
+ * Retrieves discounts based on the provided filters and pagination parameters.
+ * 
+ * @param request - The NextRequest object containing the query parameters.
+ * @returns A NextResponse object containing the fetched discounts or an error response.
+ */
 const getDiscounts = async (request: NextRequest) => {
   try {
     // Extract the filters from the query params
@@ -93,7 +104,7 @@ const getDiscounts = async (request: NextRequest) => {
     const user = await currentUser();
     // If the user is logged in, fetch private group discounts and public discounts
     if (userId && user) {
-      const supabase = await supabaseClientWithAccessToken(
+      const supabase = await supabaseClient(
         request.headers.get("supabase_jwt")
       );
       if (!supabase) {
@@ -153,4 +164,129 @@ const getDiscounts = async (request: NextRequest) => {
   }
 };
 
-export { insertDiscount, getDiscounts };
+/**
+ * Deletes a discount from the database.
+ * 
+ * @param request - The NextRequest object containing the request details.
+ * @returns A NextResponse object with the result of the deletion.
+ */
+const deleteDiscount = async (request: NextRequest) => {
+  try {
+    const { userId, getToken, orgRole } = auth();
+    const user = await currentUser();
+
+    if (userId && user) {
+      const discount_id = request.nextUrl.searchParams.get("discount_id");
+
+      // Create a Supabase client with the current user's access token
+      const token = request.headers.get("supabase_jwt");
+      if (!token) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+      }
+      const supabase = await supabaseClient(token);
+
+      const { error } = await supabase
+        .from("discounts")
+        .delete()
+        .eq("id", discount_id);
+
+      if (error) {
+        return NextResponse.json(
+          { error: "Failed to delete discount" },
+          { status: 500 }
+        );
+      } else {
+        return NextResponse.json(
+          { success: true, deleted: discount_id },
+          { status: 200 }
+        );
+      }
+    } else {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+  } catch (error) {
+    return NextResponse.json(
+      { error: "Internal Server Error" },
+      { status: 500 }
+    );
+  }
+};
+
+/**
+ * Updates a discount in the database.
+ * 
+ * @param request - The NextRequest object containing the request details.
+ * @returns A NextResponse object with the updated discount data or an error message.
+ */
+const updateDiscount = async (request: NextRequest) => {
+  try {
+    const { userId, getToken } = auth();
+    const user = await currentUser();
+    if (userId && user) {
+      // Create a Supabase client with the current user's access token
+      const token = request.headers.get("supabase_jwt");
+      if (!token) {
+        console.log("no token");
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+      }
+      const supabase = await supabaseClient(token);
+
+      const formData = await request.formData();
+      const updatedDiscount: any = {};
+
+      const discount_id = formData.get("discount_id");
+
+      if (formData.get("company")) {
+        updatedDiscount.company = formData.get("company");
+      }
+      if (formData.get("terms_and_conditions")) {
+        updatedDiscount.terms_and_conditions = formData.get(
+          "terms_and_conditions"
+        );
+      }
+      if (formData.get("company_url")) {
+        updatedDiscount.company_url = formData.get("company_url");
+      }
+      if (formData.get("discount_amount")) {
+        updatedDiscount.discount_amount = Number(
+          formData.get("discount_amount")
+        );
+      }
+      if (formData.get("public")) {
+        updatedDiscount.public =
+          formData.get("public") === "true" ? true : false;
+      }
+      if (formData.get("private_groups")) {
+        updatedDiscount.private_groups = formData.get("private_groups");
+      }
+
+      const { data, error } = await supabase
+        .from("discounts")
+        .update(updatedDiscount)
+        .eq("id", discount_id)
+        .select();
+
+      if (error) {
+        console.log(error);
+        return NextResponse.json(
+          { error: "Failed to update discount" },
+          { status: 500 }
+        );
+      } else {
+        return NextResponse.json(
+          { success: true, updated_values: data },
+          { status: 200 }
+        );
+      }
+    } else {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+  } catch (error) {
+    return NextResponse.json(
+      { error: "Internal Server Error" },
+      { status: 500 }
+    );
+  }
+};
+
+export { insertDiscount, getDiscounts, deleteDiscount, updateDiscount };
